+++ src/Utility/Stringify.php
@@ -7,6 +7,15 @@
 
 namespace QL\Panthor\Utility;
 
+/**
+ * This utility provides helper functions to combine strings.
+ *
+ * It should only be used for Symfony DI configuration, as sometimes you have scalars as services
+ * (or synthetic services) that need to be appended to other parameters (such as file paths).
+ *
+ * For example, Panthor uses this for building the full path of the twig template directory by combining
+ * the app root (@root synthetic service) to the relative template path (by default "templates").
+ */
 class Stringify
 {
     /**